+++ packages/extension-koni-ui/src/Popup/Transaction/variants/Swap/index.tsx
@@ -6,12 +6,13 @@ import { SwapError } from '@subwallet/extension-base/background/errors/SwapError
 import { ExtrinsicType, NotificationType } from '@subwallet/extension-base/background/KoniTypes';
 import { validateRecipientAddress } from '@subwallet/extension-base/core/logic-validation/recipientAddress';
 import { ActionType } from '@subwallet/extension-base/core/types';
+import { _BALANCE_CHAIN_GROUP } from '@subwallet/extension-base/services/chain-service/constants';
 import { _ChainState } from '@subwallet/extension-base/services/chain-service/types';
-import { _getAssetDecimals, _getAssetOriginChain, _getAssetSymbol, _getChainName, _getMultiChainAsset, _getOriginChainOfAsset, _isAssetFungibleToken, _isChainEvmCompatible, _isChainInfoCompatibleWithAccountInfo, _parseAssetRefKey } from '@subwallet/extension-base/services/chain-service/utils';
+import { _getAssetDecimals, _getAssetOriginChain, _getAssetSymbol, _getChainName, _getMultiChainAsset, _getOriginChainOfAsset, _isAssetFungibleToken, _isChainEvmCompatible, _isChainInfoCompatibleWithAccountInfo, _isNativeTokenBySlug, _parseAssetRefKey } from '@subwallet/extension-base/services/chain-service/utils';
 import { KyberSwapQuoteMetadata } from '@subwallet/extension-base/services/swap-service/handler/kyber-handler';
 import { DetectedGenOptimalProcessErrMsg } from '@subwallet/extension-base/services/swap-service/utils';
 import { SWTransactionResponse } from '@subwallet/extension-base/services/transaction-service/types';
-import { AccountChainType, AccountProxy, AccountProxyType, AnalyzedGroup, CommonOptimalSwapPath, ProcessType, SwapRequestResult, SwapRequestV2 } from '@subwallet/extension-base/types';
+import { AccountChainType, AccountProxy, AccountProxyType, AnalyzedGroup, BalanceType, CommonOptimalSwapPath, ProcessType, SwapRequestResult, SwapRequestV2 } from '@subwallet/extension-base/types';
 import { CHAINFLIP_SLIPPAGE, SIMPLE_SWAP_SLIPPAGE, SlippageType, SwapProviderId, SwapQuote } from '@subwallet/extension-base/types/swap';
 import { isSameAddress } from '@subwallet/extension-base/utils';
 import { getId } from '@subwallet/extension-base/utils/getId';
@@ -119,7 +120,7 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
 
   const { activeModal, inactiveAll, inactiveModal } = useContext(ModalContext);
 
-  const { accountProxies, accounts, isAllAccount } = useSelector((state) => state.accountState);
+  const { accountProxies, accounts, currentAccountProxy, isAllAccount } = useSelector((state) => state.accountState);
   const assetRegistryMap = useSelector((state) => state.assetRegistry.assetRegistry);
   const { priceMap } = useSelector((state) => state.price);
   const { chainInfoMap, chainStateMap, ledgerGenericAllowNetworks } = useSelector((root) => root.chainStore);
@@ -188,7 +189,27 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
   const chainValue = useWatchTransaction('chain', form, defaultData);
   const recipientValue = useWatchTransaction('recipient', form, defaultData);
 
-  const availableBalanceHookResult = useGetBalance(chainValue, fromValue, fromTokenSlugValue, true, ExtrinsicType.SWAP);
+  const fromAssetInfo = useMemo(() => {
+    return assetRegistryMap[fromTokenSlugValue] || undefined;
+  }, [assetRegistryMap, fromTokenSlugValue]);
+
+  const toAssetInfo = useMemo(() => {
+    return assetRegistryMap[toTokenSlugValue] || undefined;
+  }, [assetRegistryMap, toTokenSlugValue]);
+
+  const isBittensorStakedSwap = useMemo(() => {
+    return _BALANCE_CHAIN_GROUP.bittensor.includes(fromAssetInfo?.originChain) && _BALANCE_CHAIN_GROUP.bittensor.includes(toAssetInfo?.originChain);
+  }, [fromAssetInfo?.originChain, toAssetInfo?.originChain]);
+
+  const balanceType = useMemo(() => {
+    if (isBittensorStakedSwap && _isNativeTokenBySlug(fromAssetInfo?.slug)) {
+      return BalanceType.STAKING;
+    } else {
+      return undefined;
+    }
+  }, [fromAssetInfo?.slug, isBittensorStakedSwap]);
+
+  const availableBalanceHookResult = useGetBalance(chainValue, fromValue, fromTokenSlugValue, true, ExtrinsicType.SWAP, balanceType);
 
   const [swapFromFieldRenderKey, setSwapFromFieldRenderKey] = useState<string>('SwapFromField');
 
@@ -346,14 +367,6 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
 
     return tokenSelectorItems.filter((item) => destinationSlugSet.has(item.slug));
   }, [fromTokenSlugValue, tokenSelectorItems, pairMap]);
-
-  const fromAssetInfo = useMemo(() => {
-    return assetRegistryMap[fromTokenSlugValue] || undefined;
-  }, [assetRegistryMap, fromTokenSlugValue]);
-
-  const toAssetInfo = useMemo(() => {
-    return assetRegistryMap[toTokenSlugValue] || undefined;
-  }, [assetRegistryMap, toTokenSlugValue]);
 
   const destChainValue = _getAssetOriginChain(toAssetInfo);
 
@@ -758,11 +771,17 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
       return;
     }
 
-    const result = new BigN(currentFromTokenAvailableBalance.value).multipliedBy(92).dividedToIntegerBy(100).toFixed();
+    let result: string;
+
+    if (isBittensorStakedSwap) {
+      result = currentFromTokenAvailableBalance.value;
+    } else {
+      result = new BigN(currentFromTokenAvailableBalance.value).multipliedBy(92).dividedToIntegerBy(100).toFixed();
+    }
 
     onChangeAmount(result);
     setSwapFromFieldRenderKey(`SwapFromField-${Date.now()}`);
-  }, [currentFromTokenAvailableBalance, onChangeAmount]);
+  }, [currentFromTokenAvailableBalance, isBittensorStakedSwap, onChangeAmount]);
 
   const onClickHaftAmountButton = useCallback(() => {
     if (!currentFromTokenAvailableBalance) {
@@ -837,6 +856,40 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
       name: accountJsonForRecipientAutoFilled.name
     });
   }, [chainInfoMap, destChainValue, getReformatAddress, isRecipientFieldAllowed, targetAccountProxy.accountType, targetAccountProxy.accounts]);
+
+  /**
+   * Retrieve the Substrate address associated with the current account proxy, or a selected account proxy in case All Account mode.
+   *
+   * The logic follows these steps:
+   * 1. Determine the AccountProxy, correspond to `currentSelectAccountProxy`.
+   * 2. Retrieve substrate address if it's available.
+   */
+  const alternativeAddress: string | undefined = useMemo(() => {
+    if (currentAccountProxy === null) {
+      return undefined;
+    }
+
+    let currentSelectAccountProxy: AccountProxy | undefined;
+
+    if (currentAccountProxy.accountType === AccountProxyType.ALL_ACCOUNT) {
+      // If in all account mode, find the selected from account by `targetAccountProxyIdForGetBalance`
+      currentSelectAccountProxy = accountProxies.find((account) => account.id === targetAccountProxyIdForGetBalance);
+    } else {
+      currentSelectAccountProxy = currentAccountProxy;
+    }
+
+    if (!currentSelectAccountProxy) {
+      return undefined;
+    }
+
+    const substrateAccount = currentSelectAccountProxy.accounts.find((account) => account.chainType === AccountChainType.SUBSTRATE);
+
+    if (!substrateAccount) {
+      return undefined;
+    }
+
+    return substrateAccount.address;
+  }, [accountProxies, currentAccountProxy, targetAccountProxyIdForGetBalance]);
 
   useEffect(() => {
     if (fromTokenSlugValue && toTokenSlugValue && isAddressInputReady) {
@@ -928,6 +981,7 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
 
           const currentRequest: SwapRequestV2 = {
             address: fromValue,
+            alternativeAddress: alternativeAddress,
             pair: {
               slug: _parseAssetRefKey(fromTokenSlugValue, toTokenSlugValue),
               from: fromTokenSlugValue,
@@ -993,7 +1047,7 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
       sync = false;
       clearTimeout(timeout);
     };
-  }, [ErrorMessageMap, currentSlippage.slippage, form, fromAmountValue, fromTokenSlugValue, fromValue, isRecipientFieldAllowed, preferredProvider, recipientValue, toTokenSlugValue, updateSwapStates, notifyErrorMessage]);
+  }, [ErrorMessageMap, currentSlippage.slippage, form, fromAmountValue, fromTokenSlugValue, fromValue, isRecipientFieldAllowed, preferredProvider, recipientValue, toTokenSlugValue, updateSwapStates, notifyErrorMessage, alternativeAddress]);
 
   useEffect(() => {
     // eslint-disable-next-line prefer-const
@@ -1316,13 +1370,22 @@ const Component = ({ allowedChainAndExcludedTokenForTargetAccountProxy, defaultS
               <div className={'__balance-display-area'}>
                 <FreeBalance
                   address={fromValue}
+                  balanceType={balanceType}
                   chain={chainValue}
                   className={'__balance-display'}
                   extrinsicType={ExtrinsicType.SWAP}
                   hidden={!canShowAvailableBalance}
                   isSubscribe={true}
-                  label={`${t('ui.TRANSACTION.screen.Transaction.Swap.availableBalance')}`}
-                  labelTooltip={'Available balance for swap'}
+                  label={balanceType === BalanceType.STAKING ? `${t('ui.TRANSACTION.screen.Transaction.Swap.stakedBalance')}` : `${t('ui.TRANSACTION.screen.Transaction.Swap.availableBalance')}`}
+                  labelTooltip={
+                    balanceType === BalanceType.STAKING
+                      ? t('ui.TRANSACTION.screen.Transaction.Swap.stakedBalanceTooltip', {
+                        replace: {
+                          symbol: fromAssetInfo.symbol
+                        }
+                      })
+                      : t('ui.TRANSACTION.screen.Transaction.Swap.availableBalanceTooltip')
+                  }
                   tokenSlug={fromTokenSlugValue}
                 />
               </div>